Add a setSelectedPaperSize method in PageTypeSelectionPopover
The setSelectedPaperSize method takes in a PaperSize and a boolean template
argument. It sets the pageSize variables to the argument in the
PageTypeSelectionPopover and PageBackgroundController class and
enables/disables the orientation-GAction based on if the 'copy current
page size' option is passed. Lastly, it's sets the orientation on the
orientation-GAction. If the template argument is true, the combo box
selection is also updated to the fitting option.